+++ next.config.js
@@ -1,0 +1,3 @@
+const removeImports = require('next-remove-imports')();
+
+module.exports = removeImports({});
