+++ package.json
@@ -21,6 +21,7 @@
     "@toast-ui/editor": "^2.5.2",
     "@toast-ui/react-editor": "^2.5.3",
     "@types/react-cookies": "^0.1.0",
+    "@uiw/react-md-editor": "^3.3.4",
     "antd": "^4.15.5",
     "axios": "^0.21.1",
     "bcryptjs": "^2.4.3",
@@ -33,6 +34,7 @@
     "mongoose": "^5.10.18",
     "mongoose-autopopulate": "^0.11.0",
     "next": "^10.2.3",
+    "next-remove-imports": "^1.0.5",
     "normalize.css": "^8.0.1",
     "react": "^17.0.1",
     "react-cookies": "^0.1.1",
