+++ pages/editor.tsx
@@ -1,36 +1,18 @@
-import React, { LegacyRef } from 'react'
+import React from 'react'
 
-import 'codemirror/lib/codemirror.css'
-import '@toast-ui/editor/dist/toastui-editor.css'
-import { Editor } from '@toast-ui/react-editor'
-import { Button, Row, Col } from 'antd'
+import Head from 'next/head'
+import MDEditor from '@uiw/react-md-editor';
+import '@uiw/react-md-editor/dist/markdown-editor.css'
+import '@uiw/react-markdown-preview/dist/markdown.css';
 
-export default function MemoEditor() {
-  const editorRef: LegacyRef<Editor> = React.createRef()
-
-  const handleClick = () => {
-    console.log(editorRef.current?.getInstance().getHtml())
-  }
-
+export default function Home() {
   return (
-    <>
-      <Row justify="end">
-        <Col span={1}>
-          <Button danger>취소</Button>
-        </Col>
-        <Col span={1}>
-          <Button type="primary" onClick={handleClick}>
-            저장
-          </Button>
-        </Col>
-      </Row>
-      <Editor
-        previewStyle="vertical"
-        height="400px"
-        initialEditType="wysiwyg"
-        initialValue={'메모를 입력하세요'}
-        ref={editorRef}
-      />
-    </>
+    <div>
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+      </Head>
+      <MDEditor value="**Hello world!!!**" />
+    </div>
   )
 }
